Use more appropriate types

+++ examples/net/zn_pull.c
@@ -16,8 +16,8 @@
 
 void data_handler(const zn_sample_t *sample, const void *arg) {
     printf(">> [Subscription listener] Received (%.*s, %.*s)\n",
-        sample->key.len, sample->key.val,
-        sample->value.len, sample->value.val);
+        (int)sample->key.len, sample->key.val,
+        (int)sample->value.len, sample->value.val);
 }
 
 int main(int argc, char **argv) {
